==> 01-Origin.Rmd <==

---
title: "Origin"
output:
  html_document:
    df_print: paged
bibliography: [book.bib, packages.bib]

---

```{r, include=FALSE, eval =TRUE}
knitr::opts_chunk$set(echo = TRUE)
library(DiagrammeR)     #for workflow diagram
library(webshot)     #for workflow diagram
```
# Origin

* Semester Date: Spring 2021
* Class:  Capstone Portfolio (STAT 6382)
* Program: Master of Science in Data Analytics
* School: University of Houston Downtown

## Starting Point
House Prices: Advanced Regression Techniques is an ongoing Kaggle competition [@kagglehouseprices]. This "getting started" competition calls for extensive data pre-processing, feature engineering, and implementation of advanced regression techniques like random forest and gradient boosting. The data-set was made available through the Kaggle website at kaggle.com [@houseprices]. 

## Technical Skills

* Techniques: Data Cleaning, Exploratory Data Analysis, Feature Engineering, Advanced Regression Techniques, Machine Learning

* Program: R

## Problem Definition
Buying a house is a complicated decision making process for a lot of people. As Kaggle.com puts it: "Ask a home buyer to describe their dream house, and they probably won't begin with the height of the basement ceiling or the proximity to an east-west railroad" [@kagglehouseprices]. However, it is in fact, characteristics like these that influence the price of homes. Throughout this project we will explore a plethora of house features from kitchen quality to land slope, and evaluate their contribution to the sale price of homes. With some luck, or rather some data analytics skills, we will gain some insight into the real estate world of pricing residential homes. By understanding the relationship between key features and price, both buyers and sellers can optimize their price negotiations game, and make smarter financial decisions in the process of buying or selling a home. 

This project has three main goals. The first, as defined by Kaggle, is to predict the sales price value for each ID in the test set using advanced regression techniques. The second goal is to identify which features affect sale price the most. Lastly, I would like to explore which data analytics techniques lead to a better prediction in this case. 


* Which are the most relevant features that impact sales price?
* Which data pre-processing techniques improve the predictive value of the data?
* Which model performs best at predicting price?


## Data Analytics Workflow
Here \@ref(fig:Workflow) is an overview of the data analytics workflow for this project done with packages 'DiagrameR' [@diagramer] and 'webshot' [@webshot].


```{r Workflow, echo=FALSE, eval =TRUE, fig.cap="Workflow", fig.align = 'center', out.width="auto"}

grViz("
digraph boxes_and_circles {
graph [layout = dot, rankdir = LR]
node [shape = rectangle,color=AliceBlue, fontname = Helvetica,
     style = filled, fillcolor = AliceBlue, fontname=Helvetica, fontsize=14, fontcolor=SlateGray]
edge [color = SteelBlue,
     arrowsize = 1, penwidth=1.2]
rec1 [label = 'Data']
rec2 [label = 'Data Preprocessing']
rec3 [label = 'Modeling']
rec4 [label = 'Prediction']
rec5 [label = 'Validation']


rec1-> rec2-> rec3-> rec4-> rec5
}",
height = 100)

```
